Refactoring to extends some classes - Extract SessionRepositoryRequestWrapper and SessionRepositoryResponseWrapper to a public class - Created methods to override wrapRequest and wrapResponse - Changed private fields to protected to override configuration - Changed RedisSession to override RedisOperationsSessionRepository
