+++ spring-session/src/main/java/org/springframework/session/config/annotation/web/http/SpringHttpSessionConfiguration.java
@@ -15,12 +15,6 @@
  */
 package org.springframework.session.config.annotation.web.http;
 
-import java.util.ArrayList;
-import java.util.List;
-
-import javax.servlet.ServletContext;
-import javax.servlet.http.HttpSessionListener;
-
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Bean;
 import org.springframework.context.annotation.Configuration;
@@ -28,12 +22,12 @@ import org.springframework.session.ExpiringSession;
 import org.springframework.session.SessionRepository;
 import org.springframework.session.events.SessionCreatedEvent;
 import org.springframework.session.events.SessionDestroyedEvent;
-import org.springframework.session.web.http.CookieHttpSessionStrategy;
-import org.springframework.session.web.http.CookieSerializer;
-import org.springframework.session.web.http.HttpSessionStrategy;
-import org.springframework.session.web.http.MultiHttpSessionStrategy;
-import org.springframework.session.web.http.SessionEventHttpSessionListenerAdapter;
-import org.springframework.session.web.http.SessionRepositoryFilter;
+import org.springframework.session.web.http.*;
+
+import javax.servlet.ServletContext;
+import javax.servlet.http.HttpSessionListener;
+import java.util.ArrayList;
+import java.util.List;
 
 /**
  * Configures the basics for setting up Spring Session in a web environment. In
@@ -82,49 +76,54 @@ import org.springframework.session.web.http.SessionRepositoryFilter;
  */
 @Configuration
 public class SpringHttpSessionConfiguration {
+// ------------------------------ FIELDS ------------------------------
 
-	private CookieHttpSessionStrategy defaultHttpSessionStrategy = new CookieHttpSessionStrategy();
+    protected CookieHttpSessionStrategy defaultHttpSessionStrategy = new CookieHttpSessionStrategy();
 
-	private HttpSessionStrategy httpSessionStrategy = defaultHttpSessionStrategy;
+    protected HttpSessionStrategy httpSessionStrategy = defaultHttpSessionStrategy;
 
-	private List<HttpSessionListener> httpSessionListeners = new ArrayList<HttpSessionListener>();
+    protected List<HttpSessionListener> httpSessionListeners = new ArrayList<HttpSessionListener>();
 
-	private ServletContext servletContext;
+    protected ServletContext servletContext;
 
-	@Bean
-	public SessionEventHttpSessionListenerAdapter sessionEventHttpSessionListenerAdapter() {
-		return new SessionEventHttpSessionListenerAdapter(httpSessionListeners);
-	}
+// --------------------- GETTER / SETTER METHODS ---------------------
 
-	@Bean
-	public <S extends ExpiringSession> SessionRepositoryFilter<? extends ExpiringSession> springSessionRepositoryFilter(SessionRepository<S> sessionRepository) {
-		SessionRepositoryFilter<S> sessionRepositoryFilter = new SessionRepositoryFilter<S>(sessionRepository);
-		sessionRepositoryFilter.setServletContext(servletContext);
-		if(httpSessionStrategy instanceof MultiHttpSessionStrategy) {
-			sessionRepositoryFilter.setHttpSessionStrategy((MultiHttpSessionStrategy) httpSessionStrategy);
-		} else {
-			sessionRepositoryFilter.setHttpSessionStrategy(httpSessionStrategy);
-		}
-		return sessionRepositoryFilter;
-	}
+    @Autowired(required = false)
+    public void setHttpSessionListeners(List<HttpSessionListener> listeners) {
+        this.httpSessionListeners = listeners;
+    }
 
-	@Autowired(required=false)
-	public void setServletContext(ServletContext servletContext) {
-		this.servletContext = servletContext;
-	}
+    @Autowired(required = false)
+    public void setHttpSessionStrategy(HttpSessionStrategy httpSessionStrategy) {
+        this.httpSessionStrategy = httpSessionStrategy;
+    }
 
-	@Autowired(required = false)
-	public void setCookieSerializer(CookieSerializer cookieSerializer) {
-		this.defaultHttpSessionStrategy.setCookieSerializer(cookieSerializer);
-	}
+    @Autowired(required = false)
+    public void setServletContext(ServletContext servletContext) {
+        this.servletContext = servletContext;
+    }
 
-	@Autowired(required = false)
-	public void setHttpSessionStrategy(HttpSessionStrategy httpSessionStrategy) {
-		this.httpSessionStrategy = httpSessionStrategy;
-	}
+// -------------------------- OTHER METHODS --------------------------
 
-	@Autowired(required = false)
-	public void setHttpSessionListeners(List<HttpSessionListener> listeners) {
-		this.httpSessionListeners = listeners;
-	}
+    @Bean
+    public SessionEventHttpSessionListenerAdapter sessionEventHttpSessionListenerAdapter() {
+        return new SessionEventHttpSessionListenerAdapter(httpSessionListeners);
+    }
+
+    @Autowired(required = false)
+    public void setCookieSerializer(CookieSerializer cookieSerializer) {
+        this.defaultHttpSessionStrategy.setCookieSerializer(cookieSerializer);
+    }
+
+    @Bean
+    public <S extends ExpiringSession> SessionRepositoryFilter<? extends ExpiringSession> springSessionRepositoryFilter(SessionRepository<S> sessionRepository) {
+        SessionRepositoryFilter<?> sessionRepositoryFilter = new SessionRepositoryFilter<S>(sessionRepository);
+        sessionRepositoryFilter.setServletContext(servletContext);
+        if (httpSessionStrategy instanceof MultiHttpSessionStrategy) {
+            sessionRepositoryFilter.setHttpSessionStrategy((MultiHttpSessionStrategy) httpSessionStrategy);
+        } else {
+            sessionRepositoryFilter.setHttpSessionStrategy(httpSessionStrategy);
+        }
+        return sessionRepositoryFilter;
+    }
 }
