+++ spring-session/src/main/java/org/springframework/session/web/http/SessionRepositoryFilter.java
@@ -15,25 +15,17 @@
  */
 package org.springframework.session.web.http;
 
-import java.io.IOException;
-import java.util.Enumeration;
-import java.util.HashMap;
-import java.util.Map;
+import org.springframework.core.annotation.Order;
+import org.springframework.session.ExpiringSession;
+import org.springframework.session.Session;
+import org.springframework.session.SessionRepository;
 
 import javax.servlet.FilterChain;
 import javax.servlet.ServletContext;
 import javax.servlet.ServletException;
 import javax.servlet.http.HttpServletRequest;
-import javax.servlet.http.HttpServletRequestWrapper;
 import javax.servlet.http.HttpServletResponse;
-import javax.servlet.http.HttpSession;
-
-import org.apache.commons.logging.Log;
-import org.apache.commons.logging.LogFactory;
-import org.springframework.core.annotation.Order;
-import org.springframework.session.ExpiringSession;
-import org.springframework.session.Session;
-import org.springframework.session.SessionRepository;
+import java.io.IOException;
 
 /**
  * Switches the {@link javax.servlet.http.HttpSession} implementation to be backed by a {@link org.springframework.session.Session}.
@@ -63,318 +55,112 @@ import org.springframework.session.SessionRepository;
  */
 @Order(SessionRepositoryFilter.DEFAULT_ORDER)
 public class SessionRepositoryFilter<S extends ExpiringSession> extends OncePerRequestFilter {
-	private static final String SESSION_LOGGER_NAME = SessionRepositoryFilter.class.getName().concat(".SESSION_LOGGER");
 
-	private static final Log SESSION_LOGGER = LogFactory.getLog(SESSION_LOGGER_NAME);
+    public static final String SESSION_REPOSITORY_ATTR = SessionRepository.class.getName();
 
-	public static final String SESSION_REPOSITORY_ATTR = SessionRepository.class.getName();
+    public static final int DEFAULT_ORDER = Integer.MIN_VALUE + 50;
 
-	public static final int DEFAULT_ORDER = Integer.MIN_VALUE + 50;
+    protected final SessionRepository<S> sessionRepository;
 
-	private final SessionRepository<S> sessionRepository;
+    protected ServletContext servletContext;
 
-	private ServletContext servletContext;
+    protected MultiHttpSessionStrategy httpSessionStrategy = new CookieHttpSessionStrategy();
 
-	private MultiHttpSessionStrategy httpSessionStrategy = new CookieHttpSessionStrategy();
+    /**
+     * Creates a new instance
+     *
+     * @param sessionRepository the <code>SessionRepository</code> to use. Cannot be null.
+     */
+    public SessionRepositoryFilter(SessionRepository<S> sessionRepository) {
+        if (sessionRepository == null) {
+            throw new IllegalArgumentException("sessionRepository cannot be null");
+        }
+        this.sessionRepository = sessionRepository;
+    }
 
-	/**
-	 * Creates a new instance
-	 *
-	 * @param sessionRepository the <code>SessionRepository</code> to use. Cannot be null.
-	 */
-	public SessionRepositoryFilter(SessionRepository<S> sessionRepository) {
-		if(sessionRepository == null) {
-			throw new IllegalArgumentException("sessionRepository cannot be null");
-		}
-		this.sessionRepository = sessionRepository;
-	}
+    /**
+     * Sets the {@link HttpSessionStrategy} to be used. The default is a {@link CookieHttpSessionStrategy}.
+     *
+     * @param httpSessionStrategy the {@link HttpSessionStrategy} to use. Cannot be null.
+     */
+    public void setHttpSessionStrategy(HttpSessionStrategy httpSessionStrategy) {
+        if (httpSessionStrategy == null) {
+            throw new IllegalArgumentException("httpSessionStrategy cannot be null");
+        }
+        this.httpSessionStrategy = new MultiHttpSessionStrategyAdapter(httpSessionStrategy);
+    }
 
-	/**
-	 * Sets the {@link HttpSessionStrategy} to be used. The default is a {@link CookieHttpSessionStrategy}.
-	 *
-	 * @param httpSessionStrategy the {@link HttpSessionStrategy} to use. Cannot be null.
-	 */
-	public void setHttpSessionStrategy(HttpSessionStrategy httpSessionStrategy) {
-		if(httpSessionStrategy == null) {
-			throw new IllegalArgumentException("httpSessionStrategy cannot be null");
-		}
-		this.httpSessionStrategy = new MultiHttpSessionStrategyAdapter(httpSessionStrategy);
-	}
+    /**
+     * Sets the {@link MultiHttpSessionStrategy} to be used. The default is a {@link CookieHttpSessionStrategy}.
+     *
+     * @param httpSessionStrategy the {@link MultiHttpSessionStrategy} to use. Cannot be null.
+     */
+    public void setHttpSessionStrategy(MultiHttpSessionStrategy httpSessionStrategy) {
+        if (httpSessionStrategy == null) {
+            throw new IllegalArgumentException("httpSessionStrategy cannot be null");
+        }
+        this.httpSessionStrategy = httpSessionStrategy;
+    }
 
-	/**
-	 * Sets the {@link MultiHttpSessionStrategy} to be used. The default is a {@link CookieHttpSessionStrategy}.
-	 *
-	 * @param httpSessionStrategy the {@link MultiHttpSessionStrategy} to use. Cannot be null.
-	 */
-	public void setHttpSessionStrategy(MultiHttpSessionStrategy httpSessionStrategy) {
-		if(httpSessionStrategy == null) {
-			throw new IllegalArgumentException("httpSessionStrategy cannot be null");
-		}
-		this.httpSessionStrategy = httpSessionStrategy;
-	}
+    protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
+        request.setAttribute(SESSION_REPOSITORY_ATTR, sessionRepository);
 
-	protected void doFilterInternal(HttpServletRequest request, HttpServletResponse response, FilterChain filterChain) throws ServletException, IOException {
-		request.setAttribute(SESSION_REPOSITORY_ATTR, sessionRepository);
+        SessionRepositoryRequestWrapper wrappedRequest = wrapRequest(request, response);
+        SessionRepositoryResponseWrapper wrappedResponse = wrapResponse(wrappedRequest, response);
 
-		SessionRepositoryRequestWrapper wrappedRequest = new SessionRepositoryRequestWrapper(request, response, servletContext);
-		SessionRepositoryResponseWrapper wrappedResponse = new SessionRepositoryResponseWrapper(wrappedRequest,response);
+        HttpServletRequest strategyRequest = httpSessionStrategy.wrapRequest(wrappedRequest, wrappedResponse);
+        HttpServletResponse strategyResponse = httpSessionStrategy.wrapResponse(wrappedRequest, wrappedResponse);
 
-		HttpServletRequest strategyRequest = httpSessionStrategy.wrapRequest(wrappedRequest, wrappedResponse);
-		HttpServletResponse strategyResponse = httpSessionStrategy.wrapResponse(wrappedRequest, wrappedResponse);
+        try {
+            filterChain.doFilter(strategyRequest, strategyResponse);
+        } finally {
+            wrappedRequest.commitSession();
+        }
+    }
 
-		try {
-			filterChain.doFilter(strategyRequest, strategyResponse);
-		} finally {
-			wrappedRequest.commitSession();
-		}
-	}
+    protected SessionRepositoryRequestWrapper wrapRequest(HttpServletRequest request, HttpServletResponse response) {
+        return new SessionRepositoryRequestWrapper(request, response, servletContext, sessionRepository, httpSessionStrategy);
+    }
 
-	public void setServletContext(ServletContext servletContext) {
-		this.servletContext = servletContext;
-	}
+    protected SessionRepositoryResponseWrapper wrapResponse(SessionRepositoryRequestWrapper request, HttpServletResponse response) {
+        return new SessionRepositoryResponseWrapper(request, response);
+    }
 
-	/**
-	 * Allows ensuring that the session is saved if the response is committed.
-	 *
-	 * @author Rob Winch
-	 * @since 1.0
-	 */
-	private final class SessionRepositoryResponseWrapper extends OnCommittedResponseWrapper {
 
-		private final SessionRepositoryRequestWrapper request;
+    public void setServletContext(ServletContext servletContext) {
+        this.servletContext = servletContext;
+    }
 
-		/**
-		 * @param response the response to be wrapped
-		 */
-		public SessionRepositoryResponseWrapper(SessionRepositoryRequestWrapper request, HttpServletResponse response) {
-			super(response);
-			if(request == null) {
-				throw new IllegalArgumentException("request cannot be null");
-			}
-			this.request = request;
-		}
 
-		@Override
-		protected void onResponseCommitted() {
-			request.commitSession();
-		}
-	}
+    static class MultiHttpSessionStrategyAdapter implements MultiHttpSessionStrategy {
+        private HttpSessionStrategy delegate;
 
-	/**
-	 * A {@link javax.servlet.http.HttpServletRequest} that retrieves the {@link javax.servlet.http.HttpSession} using a
-	 * {@link org.springframework.session.SessionRepository}.
-	 *
-	 * @author Rob Winch
-	 * @since 1.0
-	 */
-	private final class SessionRepositoryRequestWrapper extends HttpServletRequestWrapper {
-		private final String CURRENT_SESSION_ATTR = HttpServletRequestWrapper.class.getName();
-		private Boolean requestedSessionIdValid;
-		private boolean requestedSessionInvalidated;
-		private final HttpServletResponse response;
-		private final ServletContext servletContext;
+        public MultiHttpSessionStrategyAdapter(HttpSessionStrategy delegate) {
+            this.delegate = delegate;
+        }
 
-		private SessionRepositoryRequestWrapper(HttpServletRequest request, HttpServletResponse response, ServletContext servletContext) {
-			super(request);
-			this.response = response;
-			this.servletContext = servletContext;
-		}
+        public String getRequestedSessionId(HttpServletRequest request) {
+            return delegate.getRequestedSessionId(request);
+        }
 
-		/**
-		 * Uses the HttpSessionStrategy to write the session id tot he response and persist the Session.
-		 */
-		private void commitSession() {
-			HttpSessionWrapper wrappedSession = getCurrentSession();
-			if(wrappedSession == null) {
-				if(isInvalidateClientSession()) {
-					httpSessionStrategy.onInvalidateSession(this, response);
-				}
-			} else {
-				S session = wrappedSession.getSession();
-				sessionRepository.save(session);
-				if(!isRequestedSessionIdValid() || !session.getId().equals(getRequestedSessionId())) {
-					httpSessionStrategy.onNewSession(session, this, response);
-				}
-			}
-		}
+        public void onNewSession(Session session, HttpServletRequest request,
+                                 HttpServletResponse response) {
+            delegate.onNewSession(session, request, response);
+        }
 
-		@SuppressWarnings("unchecked")
-		private HttpSessionWrapper getCurrentSession() {
-			return (HttpSessionWrapper) getAttribute(CURRENT_SESSION_ATTR);
-		}
+        public void onInvalidateSession(HttpServletRequest request,
+                                        HttpServletResponse response) {
+            delegate.onInvalidateSession(request, response);
+        }
 
-		private void setCurrentSession(HttpSessionWrapper currentSession) {
-			if(currentSession == null) {
-				removeAttribute(CURRENT_SESSION_ATTR);
-			} else {
-				setAttribute(CURRENT_SESSION_ATTR, currentSession);
-			}
-		}
+        public HttpServletRequest wrapRequest(HttpServletRequest request,
+                                              HttpServletResponse response) {
+            return request;
+        }
 
-		@SuppressWarnings("unused")
-		public String changeSessionId() {
-			HttpSession session = getSession(false);
-
-			if(session == null) {
-				throw new IllegalStateException("Cannot change session ID. There is no session associated with this request.");
-			}
-
-			// eagerly get session attributes in case implementation lazily loads them
-			Map<String,Object> attrs = new HashMap<String,Object>();
-			Enumeration<String> iAttrNames = session.getAttributeNames();
-			while(iAttrNames.hasMoreElements()) {
-				String attrName = iAttrNames.nextElement();
-				Object value = session.getAttribute(attrName);
-
-				attrs.put(attrName, value);
-			}
-
-			sessionRepository.delete(session.getId());
-			HttpSessionWrapper original = getCurrentSession();
-			setCurrentSession(null);
-
-			HttpSessionWrapper newSession = getSession();
-			original.setSession(newSession.getSession());
-
-			newSession.setMaxInactiveInterval(session.getMaxInactiveInterval());
-			for(Map.Entry<String, Object> attr : attrs.entrySet()) {
-				String attrName = attr.getKey();
-				Object attrValue = attr.getValue();
-				newSession.setAttribute(attrName, attrValue);
-			}
-			return newSession.getId();
-		}
-
-		public boolean isRequestedSessionIdValid() {
-			if(requestedSessionIdValid == null) {
-				String sessionId = getRequestedSessionId();
-				S session = sessionId == null ? null : getSession(sessionId);
-				return isRequestedSessionIdValid(session);
-			}
-
-			return requestedSessionIdValid;
-		}
-
-		private boolean isRequestedSessionIdValid(S session) {
-			if(requestedSessionIdValid == null) {
-				requestedSessionIdValid = session != null;
-			}
-			return requestedSessionIdValid;
-		}
-
-		private boolean isInvalidateClientSession() {
-			return getCurrentSession() == null && requestedSessionInvalidated;
-		}
-
-		private S getSession(String sessionId) {
-			S session = sessionRepository.getSession(sessionId);
-			if(session == null) {
-				return null;
-			}
-			session.setLastAccessedTime(System.currentTimeMillis());
-			return session;
-		}
-
-		@Override
-		public HttpSessionWrapper getSession(boolean create) {
-			HttpSessionWrapper currentSession = getCurrentSession();
-			if(currentSession != null) {
-				return currentSession;
-			}
-			String requestedSessionId = getRequestedSessionId();
-			if(requestedSessionId != null) {
-				S session = getSession(requestedSessionId);
-				if(session != null) {
-					this.requestedSessionIdValid = true;
-					currentSession = new HttpSessionWrapper(session, getServletContext());
-					currentSession.setNew(false);
-					setCurrentSession(currentSession);
-					return currentSession;
-				}
-			}
-			if(!create) {
-				return null;
-			}
-			if(SESSION_LOGGER.isDebugEnabled()) {
-				SESSION_LOGGER
-						.debug("A new session was created. To help you troubleshoot where the session was created we provided a StackTrace (this is not an error). You can prevent this from appearing by disabling DEBUG logging for "
-								+ SESSION_LOGGER_NAME, new RuntimeException("For debugging purposes only (not an error)"));
-			}
-			S session = sessionRepository.createSession();
-			session.setLastAccessedTime(System.currentTimeMillis());
-			currentSession = new HttpSessionWrapper(session, getServletContext());
-			setCurrentSession(currentSession);
-			return currentSession;
-		}
-
-		public ServletContext getServletContext() {
-			if(servletContext != null) {
-				return servletContext;
-			}
-			// Servlet 3.0+
-			return super.getServletContext();
-		}
-
-		@Override
-		public HttpSessionWrapper getSession() {
-			return getSession(true);
-		}
-
-		@Override
-		public String getRequestedSessionId() {
-			return httpSessionStrategy.getRequestedSessionId(this);
-		}
-
-		/**
-		 * Allows creating an HttpSession from a Session instance.
-		 *
-		 * @author Rob Winch
-		 * @since 1.0
-		 */
-		private final class HttpSessionWrapper extends ExpiringSessionHttpSession<S> {
-
-			public HttpSessionWrapper(S session, ServletContext servletContext) {
-				super(session, servletContext);
-			}
-
-			public void invalidate() {
-				super.invalidate();
-				requestedSessionInvalidated = true;
-				setCurrentSession(null);
-				sessionRepository.delete(getId());
-			}
-		}
-	}
-
-	static class MultiHttpSessionStrategyAdapter implements MultiHttpSessionStrategy {
-		private HttpSessionStrategy delegate;
-
-		public MultiHttpSessionStrategyAdapter(HttpSessionStrategy delegate) {
-			this.delegate = delegate;
-		}
-
-		public String getRequestedSessionId(HttpServletRequest request) {
-			return delegate.getRequestedSessionId(request);
-		}
-
-		public void onNewSession(Session session, HttpServletRequest request,
-				HttpServletResponse response) {
-			delegate.onNewSession(session, request, response);
-		}
-
-		public void onInvalidateSession(HttpServletRequest request,
-				HttpServletResponse response) {
-			delegate.onInvalidateSession(request, response);
-		}
-
-		public HttpServletRequest wrapRequest(HttpServletRequest request,
-				HttpServletResponse response) {
-			return request;
-		}
-
-		public HttpServletResponse wrapResponse(HttpServletRequest request,
-				HttpServletResponse response) {
-			return response;
-		}
-	}
+        public HttpServletResponse wrapResponse(HttpServletRequest request,
+                                                HttpServletResponse response) {
+            return response;
+        }
+    }
 }
